Remove 1Password from all our input texts
There might be a place where we want 1Password to allow for showing the
autofill dialog, but any place I found that was showing it it was more
annoying.

For example, it shows over the Contact button in the Address input,
because it wants to fill it with my street address.