fix: Allow people to mint/burn/send "fake MFX"
The current flow looks for any mention of mfx in the denomination so
even tokens that dont want to fake MFX but use MFX in the name would
not be mint or burn-able. This fixes it.

Real umfx tokens should still not be able to be burnt or minted. The
verified badge will not show up for fake MFX tokens so this should
not be confusing to the user.

Fixes #249