fix: themes are now properly applied (#330)
Although there could be less theme pickers, this should fix themes all
around the DOM as it was querying for the theme-controller input value
and was confusing the mobileNav and sidebarNav checkboxes. There are
still both in mobile but at least they'll have the same value.

Included here are tests for useLocalStorage, better local storage and
a much simpler useTheme.