fix: Move execution of fee estimation within the Tx bounds (#332)
This allows for isSigning to be set right away, before proceeding
to estimating fees, which can take a while. Thus every component
reacting to isSigning (and showing the three bouncing dots) will
be disabled right away, without waiting for server requests to
finish.

Also moved the validation schema for sendForm to its own schema
file and import. I did not want to touch ibcSendForm, but the
idea would be to reuse the schema there too. Will do in a follow
up PR.

Finally, fixed some internals with AmountInput and the way it
interacts with schemas and Formik. This new version returns only
actual number (or undefined) instead of a string that needs
parsing. This makes it easier to use in most cases too.