fix: reintroduce sign modal ID and distinguish between inner SignModal
The issue is that the Group > Update flow has two <SignModal /> components
and they both show up at the same time, but don't interact properly
together. This leads to improper behaviour where keys and clicks being
suppressed, while moving focus between the front and back dialog closes
both. Using an ID is not pretty but is the simplest solution to this problem.

This also simplified the signing dialog to bring it inline with ModalDialog
around the application, as well as removing an issue with the ref count of
isSigning.

Fixes #367